Update readme file to reflect recent enhancements
Needed an update, as this polyfill's been extended to cover shortcomings
beyond those of simply IE8-9 (patching buggy handling of add/remove, and
adding support for forced toggling). Previously, it was just a crude way
to enable support for DOMTokenLists in antique Explorer, but was written
when I was unaware of bugs in more recent versions.